Completely refactor tunes; split into separate commands entirely
